+++ src/zope/app/applicationcontrol/zopeversion.py
@@ -16,7 +16,6 @@
 __docformat__ = 'restructuredtext'
 
 import os
-import re
 import subprocess
 
 from zope.applicationcontrol.interfaces import IZopeVersion
@@ -24,8 +23,6 @@ from zope.interface import implementer
 
 @implementer(IZopeVersion)
 class ZopeVersion(object):
-
-    __tags = re.compile(r'/(tags|branches)/([^/]+)/')
 
     def __init__(self, path=None):
         if path is None:
@@ -48,58 +45,13 @@ class ZopeVersion(object):
 
         self.result = "Development/Unknown"
 
-        # is this a SVN checkout?
-        svndir = os.path.join(self.path, ".svn")
-        if os.path.isdir(svndir):
-            self.__setSVNVersion(svndir)
-        else:
-            # try to get official Zope release information
-            versionfile = os.path.join(self.path, "version.txt")
-            if os.path.isfile(versionfile):
-                with open(versionfile) as f:
-                    self.result = f.readline().strip() or self.result
+        # try to get official Zope release information
+        versionfile = os.path.join(self.path, "version.txt")
+        if os.path.isfile(versionfile):
+            with open(versionfile) as f:
+                self.result = f.readline().strip() or self.result
 
         return self.result
 
-    def _getSVNInfoOutput(self):
-        try:
-            env = os.environ.copy()
-            env['LANG'] = env['LC_ALL'] = env['LC_MESSAGES'] = 'C'
-            proc = subprocess.Popen('svn info "%s"' % self.path,
-                                    shell=True, stdout=subprocess.PIPE, env=env)
-        except OSError:
-            pass
-        else:
-            if proc.wait() == 0:
-                return proc.stdout
-        return None
-
-    def __setSVNVersion(self, svndir):
-        output = self._getSVNInfoOutput()
-        if not output:
-            return
-
-        info = {}
-        for line in output:
-            parts = line.rstrip().split(": ", 1)
-            if len(parts) == 2:
-                key, value = parts
-                info[key] = value
-
-        revision = info.get("Revision", "")
-        url = info.get("URL", "")
-
-        if revision and url:
-            match = self.__tags.search(url)
-            if match is None:
-                tag = ""
-            else:
-                type, value = match.groups()
-                if type == "tags":
-                    tag = "/Tag: %s" % value
-                elif type == "branches":
-                    tag = "/Branch: %s" % value
-            self.result = ("Development/Revision: %s%s"
-                          % (revision, tag))
 
 ZopeVersionUtility = ZopeVersion()